% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substringRight.r
\name{substrRight}
\alias{substrRight}
\title{Takes a substring of a character string counting from the RHS}
\usage{
substrRight(x, start, stop, reverse)
}
\arguments{
\item{x}{a character vector}

\item{start}{integer. The first element to be replaced (counting from the RHS).}

\item{stop}{integer. The last element to be replaced (counting from the RHS).}

\item{reverse}{logical. Reverse the ordering of characters. Defaults to FALSE.}
}
\description{
This function allows the user to extract a substring of a character string whilst containing the start
and stop values from the RHS of the string. The user is also given the option of reversing the extracted string before
returning the extracted value.
}
\examples{
x <- '91726352EXTRACT ME2345'
substrRight(x = x, start = 5, stop = 14)

x <- '15679NODNOL'
substrRight(x = x, start = 1, stop = 6, reverse = TRUE)

}
